+++ examples/smart_id_client.rs
@@ -6,7 +6,11 @@ use anyhow::Result;
 use base64::Engine;
 use tracing_subscriber::fmt::format::FmtSpan;
 use tracing_subscriber::fmt::SubscriberBuilder;
-use smart_id_rust_client::client::v3::smart_id_client::SmartIdClientV3;
+use smart_id_rust_client::client::smart_id_client::SmartIdClientV3;
+use smart_id_rust_client::models::authentication_session::AuthenticationRequest;
+use smart_id_rust_client::models::dynamic_link::DynamicLinkType;
+use smart_id_rust_client::models::interaction::Interaction;
+use smart_id_rust_client::models::signature_session::SignatureRequest;
 use smart_id_rust_client::models::v3::authentication_session::AuthenticationRequest;
 use smart_id_rust_client::models::v3::dynamic_link::DynamicLinkType;
 use smart_id_rust_client::models::v3::interaction::Interaction;
