+++ src/client/mod.rs
@@ -1,3 +1,2 @@
 mod reqwest_generic;
-pub mod v3;
-pub mod v2;
+pub mod smart_id_client;
