--- src/client/v2/mod.rs
@@ -1,1 +1,0 @@
-pub mod smart_id_connector;
