--- src/client/v3/mod.rs
@@ -1,1 +1,0 @@
-pub mod smart_id_client;
