--- src/models/v2/mod.rs
@@ -1,4 +1,0 @@
-pub mod requests;
-pub mod responses;
-pub mod common;
-pub mod session;
