--- src/models/v3/mod.rs
@@ -1,8 +1,0 @@
-pub mod session_status;
-pub mod common;
-pub mod certificate_choice_session;
-pub mod signature_session;
-pub mod authentication_session;
-pub mod dynamic_link;
-pub mod signature;
-pub mod interaction;
