+++ src/utils/mod.rs
@@ -1,2 +1,1 @@
 pub mod sec_x509;
-pub mod verification;
